:art: Use steady_clock instead of high_resolution_clock
`high_resolution_clock` sounds good but doesn't deliver: it's not well-defined
by the standard. In libc++ and MS STL it is an alias for `steady_clock`; however
in libstdc++ it is an alias for `system_clock` (which is not steady).

As a rule of thumb, `high_resolution_clock` should never be used. For a
stopwatch, use `steady_clock`. For a wall time, use `system_clock`.

https://stackoverflow.com/a/38253266