Update flow builder to use C++20 algorithms
The change to `flow::generic_builder::hasNoIncomingEdges()` resulted in
a build failure in GCC, but success in Clang. This was tracked down to
the `begin()` and `end()` iterators some how resulting in a different
type in GCC. This then caused GCC's implementation to fail when it
attempted to subtract the iterators.

This was fixed by always using the `std::{c}begin()` method to generate
both the begin and end iterators. Other occurrences of the generation of
end iterators in a similar way were identified and corrected in
`Vector`.