:new: Add resize_and_overwrite for vector
There are several use cases where `resize_and_overwrite` is a useful function
for message data. With static capacity, the interface is not quite the same as
on `std::string`, hence it's a free function. But `resize_and_overwrite` is
still a good name, because that's what it does.

Also, simplify constructors for message_base.