:construction_worker: Turn off unit tests with clang-14 and libstdc++
There seems to be a known issue with consteval in clang-14 that causes a compile
error with libstdc++. I'm not sure why it didn't show up until now.

It manifests with the error: `call to consteval function
'std::chrono::hh_mm_ss::_S_fractional_width' is not a constant expression` when
compiling anything including chrono.