:art: RLE indices no longer materialise `bitset`s
The previous interim implementation of RLE encoded message indices would
decode data into fully materialised `bitset`s in order to fit into the
existing infrastructure easily.

This commit now eliminates the need to materialise a `bitset` to compute
which callbacks are to be executed for a given message.

The `rle_intersect` type is a proxy object that collects together
multiple `rle_decoder`s created from the indexed RLE data during message
matching. This proxy object can then be used to virtually combine the
run length encoded bitset data and enumerate the resulting `1` bits.

It is still possible to materialise the `bitset`s if needed.