cmake: use ${PROJECT_NAME}_IS_TOP_LEVEL instead of PROJECT_IS_TOP_LEVEL
I use FetchContent_Declare() and FetchContent_MakeAvailable()
to pull in the cib dependency to an ESP-IDF project (among others).
Strangely, at least with cmake version 3.28.3, PROJECT_IS_TOP_LEVEL
evaluates to true in such scenarios. The outcome of the mixup is a
completely broken build. Using ${PROJECT_NAME}_IS_TOP_LEVEL solves the
problem, as it produces the expected behavior.